add otsu method

+++ preProcessing/Histogram_Lib/histogram.h
@@ -9,7 +9,7 @@
 #include "SDL2/SDL_image.h"
 #include "../SDL_Function/sdlFunction.h"
 
-int* findHistogram(SDL_Surface* img);
-SDL_Surface* createHistogramImg(int* histogram);
+double* findHistogram(SDL_Surface* img);
+SDL_Surface* createHistogramImg(double* histogram);
 
 #endif
